Final-Exam/requirements.txt
flask
gunicorn
Flask-Failsafe
mysql-connector-python
scrypt